+++ src/main/java/pw/lemmmy/ts3protocol/packets/ack/PacketPong.java
@@ -12,7 +12,7 @@ public class PacketPong extends PacketAcknowledgement {
 		super();
 	}
 	
-	public PacketPong(short packetID) {
+	public PacketPong(int packetID) {
 		super(packetID);
 	}
 }
